+++ src/containers/PlayerContainer/PlayerContainer.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useLayoutEffect, useMemo, useState } from 'react';
 
 import { useWatchHistoryListener } from '#src/hooks/useWatchHistoryListener';
 import type { PlaylistItem } from '#types/playlist';
@@ -9,6 +9,11 @@ import Player from '#components/Player/Player';
 import type { JWPlayer } from '#types/jwplayer';
 import { useWatchHistoryStore } from '#src/stores/WatchHistoryStore';
 import { VideoProgressMinMax } from '#src/config';
+import useContentProtection from '#src/hooks/useContentProtection';
+import { getMediaById } from '#src/services/api.service';
+import LoadingOverlay from '#components/LoadingOverlay/LoadingOverlay';
+import useEventCallback from '#src/hooks/useEventCallback';
+import { logDev } from '#src/utils/common';
 import { useSettingsStore } from '#src/stores/SettingsStore';
 
 type Props = {
@@ -44,6 +49,7 @@ const PlayerContainer: React.FC<Props> = ({
   const { features } = useConfigStore((s) => s.config);
   const continueWatchingList = features?.continueWatchingList;
   const watchHistoryEnabled = !!continueWatchingList;
+  const { data: playableItem, isLoading } = useContentProtection('media', item.mediaid, (token, drmPolicyId) => getMediaById(item.mediaid, token, drmPolicyId));
 
   // state
   const [playerInstance, setPlayerInstance] = useState<JWPlayer>();
@@ -69,7 +75,13 @@ const PlayerContainer: React.FC<Props> = ({
       return null;
     }
 
-    return playerInstance.getPosition() / item.duration;
+    // this call may fail when the player is being removed due to a race condition
+    try {
+      return playerInstance.getPosition() / item.duration;
+    } catch (error: unknown) {
+      logDev('Error caught while calling `getPosition`');
+      return null;
+    }
   }, [playerInstance, item.duration]);
 
   useWatchHistoryListener(() => (watchHistoryEnabled ? saveItem(item, getProgress()) : null));
@@ -88,11 +100,11 @@ const PlayerContainer: React.FC<Props> = ({
     }
   }, [liveFromBeginning, playerInstance]);
 
-  const handleWatchHistory = useCallback(() => {
+  const handleWatchHistory = useEventCallback(() => {
     if (watchHistoryEnabled) {
       saveItem(item, getProgress());
     }
-  }, [watchHistoryEnabled, getProgress, item]);
+  });
 
   const handlePause = useCallback(() => {
     handleWatchHistory();
@@ -106,23 +118,29 @@ const PlayerContainer: React.FC<Props> = ({
 
   const handleRemove = useCallback(() => {
     handleWatchHistory();
+    setPlayerInstance(undefined);
   }, [handleWatchHistory]);
 
   const handlePlaylistItemCallback = usePlaylistItemCallback(liveStartDateTime, liveEndDateTime);
 
   // Effects
-  useEffect(() => {
+
+  // use layout effect to prevent a JWPlayer error when the instance has been removed while loading the entitlement
+  useLayoutEffect(() => {
     // save watch history when the item changes
     return () => handleWatchHistory();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [item]);
+  }, [handleWatchHistory, item]);
+
+  if (!playableItem || isLoading) {
+    return <LoadingOverlay inline />;
+  }
 
   return (
     <Player
       playerId={playerId}
       playerKey={playerKey}
       feedId={feedId}
-      item={item}
+      item={playableItem}
       onReady={handleReady}
       onFirstFrame={handleFirstFrame}
       onPlay={onPlay}