+++ src/hooks/useMedia.ts
@@ -1,8 +1,12 @@
-import useContentProtection from '#src/hooks/useContentProtection';
+import { UseBaseQueryResult, useQuery } from 'react-query';
+
 import { getMediaById } from '#src/services/api.service';
+import type { PlaylistItem } from '#types/playlist';
 
-export default function useMedia(mediaId: string, enabled: boolean = true) {
-  const callback = (token?: string, drmPolicyId?: string) => getMediaById(mediaId, token, drmPolicyId);
+export type UseMediaResult<TData = PlaylistItem, TError = unknown> = UseBaseQueryResult<TData, TError>;
 
-  return useContentProtection('media', mediaId, callback, {}, enabled);
+export default function useMedia(mediaId: string, enabled: boolean = true): UseMediaResult {
+  return useQuery(['media', mediaId], () => getMediaById(mediaId), {
+    enabled: !!mediaId && enabled,
+  });
 }