+++ src/services/inplayer.account.service.ts
@@ -34,7 +34,7 @@ enum InPlayerEnv {
 }
 
 export const setEnvironment = (config: Config) => {
-  const env: string = config.integrations?.jwp?.useSandbox ? InPlayerEnv.Daily : InPlayerEnv.Production;
+  const env: string = config.integrations?.jwp?.useSandbox ? InPlayerEnv.Development : InPlayerEnv.Production;
   InPlayer.setConfig(env as Env);
 };
 
@@ -272,18 +272,16 @@ export const subscribeToNotifications = async (uuid: string = '', onMessage: (pa
 export const updatePersonalShelves: UpdatePersonalShelves = async (payload) => {
   const { favorites, history } = payload.externalData;
   const externalData = await getCustomerExternalData();
+  const currentWatchHistoryIds = externalData?.history?.map((e) => e.mediaid);
   const currentFavoriteIds = externalData?.favorites?.map((e) => e.mediaid);
   const payloadFavoriteIds = favorites?.map((e) => e.mediaid);
 
   try {
     history.forEach(async (history) => {
-      if (externalData?.history?.length) {
-        externalData?.history?.forEach(async (historyStore) => {
-          if (historyStore.mediaid === history.mediaid && historyStore.progress !== history.progress) {
-            await InPlayer.Account.updateWatchHistory(history.mediaid, history.progress);
-          }
-        });
-      } else {
+      if (
+        !currentWatchHistoryIds?.includes(history.mediaid) ||
+        externalData?.history?.some((e) => e.mediaid == history.mediaid && e.progress != history.progress)
+      ) {
         await InPlayer.Account.updateWatchHistory(history.mediaid, history.progress);
       }
     });